+++ CHANGES.rst
@@ -10,7 +10,7 @@ Changelog of lizard-wms
 - URLs in the popup are now opened in a new window. Customers keep asking
   about this...
 
-- Improved FilterPage's breadcrumbs.
+- Improved FilterPage's breadcrumbs. We need lizard-maptree >= 1.9 for this.
 
 
 1.21 (2013-02-25)